W-20192570: Madhur/async loading (#6658)
* chore: wip

* fix: move loading of file system async

* fix: kni[

* fix: share more code

* fix: minor clean up

* fix: fix exports and ut

* fix: pr feedback

* fix: reference issues

* chore: normalize all

* Revert "chore: normalize all"

This reverts commit b27dfaa6d903b1182c902c85b31d5b418ddd7a0e.

* W-20150957 fix: replace jsonc-parser lib (#6665)

* fix: replace jsonc-parser lib

* fix: pr feedback

* fix: uncomment the line

* W-19345728 : Madhur/fix windows ut (#6666)

* fix: try unixify

* fix: unixify

* fix: unixify

* fix: other changes

* fix: remove unnecessary normalize

* fix: try actual path for windows

* fix: wip

* fix: wip

* fix: add logs

* chore: wip

* chore: add more logs

* chore: add more logs

* chore: wip

* chore: log and throw

* chore: log and throw

* fix: remove log supression

* fix: add log at start

* fix: process stdout

* fix: try diff mocking

* fix: adding mock logs for windows

* chore: test test

* chore: test test

* chore: more logs

* chore: more logs

* chore: more logs

* chore: more logs

* fix: update method

* chore: wip

* chore: wip

* chore: wip

* chore: wip

* chore: wip

* chore: wip

* chore: wip

* fix: path instead of uri

* fix: normalize some more

* chore: more logs

* fix: update directory listing

* chore: logs

* chore: directory listing infer

* fix: remove debug logs

* fix: verify if windows still good

* fix: add server to client notification for ease

* fix: try bundling fixes

* fix: remove unnecessary test

* fix: some clean up and wait in tests for server to be ready

* fix: use windows/logMessage and add more pause

* fix: remove unnecessary changes

* fix: move normalizePath to entry points

* fix: windows fix

* fix: normalize path after joining

* fix: windows path normalize

* fix: don't use nullish

* fix: add checks

* fix: getWorkspaceRoot

* fix: add logs

* fix: add logs

* fix: more logs to find rc

* chore: more logs

* chore: more logs

* chore: more logs hopefully last time

* chore: more and more logs

* chore: more and more logs

* fix: more logs

* fix: normalize test utils

* fix: context normalize

* fix: more fixes

* fix: remove debug logs

* fix: lwc context

* fix: normalize path.join

* fix: normalize test paths

* fix: namesspaceUtils

* fix: more normalized

* fix: more normalize

* fix: try a retry

* fix: self review

* fix: pr feedback

* fix: windows fix

* fix: add reload

* fix: making fsp expect normalize paths

* fix: wireit for new packages

* fix: wire it for lsp-common

* fix: remove generated tern files from git tracking

* chore: refactoring and ci for W-19345728 (#6710)

* chore: refactors and wireit setup

* chore: refactoring and linter

* fix: vscode package

* fix: adding more buffer

* fix: pass in workspaceType from client

* fix: revert linux timeput

* fix: remove unnecessary changes

* fix: try adding retries for delays

* fix: build fix

---------


---------


---------


---------

+++ scripts/vsce-bundled-extension.ts
@@ -41,8 +41,12 @@ if (!packagingConfig) {
 for (let i = 0; i < packagingConfig.assets.length; i++) {
   const asset = packagingConfig.assets[i];
   const from = `${extensionDirectory}/${asset}`;
-  logger(`copying ${from}`);
-  cpSync(from, `${directoryToConstruct}/${asset}`, { recursive: true });
+  if (existsSync(from)) {
+    logger(`copying ${from}`);
+    cpSync(from, `${directoryToConstruct}/${asset}`, { recursive: true });
+  } else {
+    logger(`skipping ${from} (does not exist - may be generated during build)`);
+  }
 }
 
 const newPackage = {
